pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forestid"
version = "0.1.0"
description = "Training and evaluating models for environment characterization"
readme = "README.md"
requires-python = ">=3.10"
# Project dependencies
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.23.0,<2",
    "dash>=2.10.0",
    "dash-core-components",
    "dash-html-components",
    "dash_bootstrap_components",
    "python-dotenv",
    "pyyaml",
    "tqdm",
    "torch @ https://download.pytorch.org/whl/cu118/torch-2.1.0%2Bcu118-cp311-cp311-linux_x86_64.whl",
    "torchvision @ https://download.pytorch.org/whl/cu118/torchvision-0.16.0%2Bcu118-cp311-cp311-linux_x86_64.whl",
    "torchaudio @ https://download.pytorch.org/whl/cu118/torchaudio-2.1.0%2Bcu118-cp311-cp311-linux_x86_64.whl"
]


[tool.setuptools]
packages = ["forestid"]

[project.optional-dependencies]
dev = [
    "scikit-learn",
    "matplotlib",
    "seaborn",
    "jupyter",
    "ipykernel",
    "pytest",
    "flake8",
    "black",
]

# Black formatter configuration
[tool.black]
line-length = 88
target-version = ['py38']

# Ruff linter configuration
[tool.ruff]
line-length = 88
target-version = "py38"